+++ Project_K/Util/Collision.cpp
@@ -20,15 +20,15 @@ bool Collision::IsColl1()
 		return false;
 	}
 
-	int PlayerLeft = m_Player[0]->GetPos().x + m_Player[0]->GetSizeLeft();
-	int PlayerTop = m_Player[0]->GetPos().y + m_Player[0]->GetSizeTop();
-	int PlayerRight = m_Player[0]->GetPos().x + m_Player[0]->GetSizeRight();
-	int PlayerBottom = m_Player[0]->GetPos().y + m_Player[0]->GetSizeBottom();
+	int PlayerLeft = static_cast<int>(m_Player[0]->GetPos().x) + m_Player[0]->GetSizeLeft();
+	int PlayerTop = static_cast<int>(m_Player[0]->GetPos().y) + m_Player[0]->GetSizeTop();
+	int PlayerRight = static_cast<int>(m_Player[0]->GetPos().x) + m_Player[0]->GetSizeRight();
+	int PlayerBottom = static_cast<int>(m_Player[0]->GetPos().y) + m_Player[0]->GetSizeBottom();
 
-	int PlayerAtkLeft = m_Player[1]->GetPos().x + m_Player[1]->GetAttackSizeLeft();
-	int PlayerAtkTop = m_Player[1]->GetPos().y + m_Player[1]->GetAttackSizeTop();
-	int PlayerAtkRight = m_Player[1]->GetPos().x + m_Player[1]->GetAttackSizeRight();
-	int PlayerAtkBottom = m_Player[1]->GetPos().y + m_Player[1]->GetAttackSizeBottom();
+	int PlayerAtkLeft = static_cast<int>(m_Player[1]->GetPos().x) + m_Player[1]->GetAttackSizeLeft();
+	int PlayerAtkTop = static_cast<int>(m_Player[1]->GetPos().y) + m_Player[1]->GetAttackSizeTop();
+	int PlayerAtkRight = static_cast<int>(m_Player[1]->GetPos().x) + m_Player[1]->GetAttackSizeRight();
+	int PlayerAtkBottom = static_cast<int>(m_Player[1]->GetPos().y) + m_Player[1]->GetAttackSizeBottom();
 
 	if (PlayerLeft > PlayerAtkRight) return false;
 	if (PlayerRight < PlayerAtkLeft) return false;
@@ -47,15 +47,15 @@ bool Collision::IsColl2()
 		return false;
 	}
 
-	int PlayerLeft = m_Player[1]->GetPos().x + m_Player[1]->GetSizeLeft();
-	int PlayerTop = m_Player[1]->GetPos().y + m_Player[1]->GetSizeTop();
-	int PlayerRight = m_Player[1]->GetPos().x + m_Player[1]->GetSizeRight();
-	int PlayerBottom = m_Player[1]->GetPos().y + m_Player[1]->GetSizeBottom();
+	int PlayerLeft = static_cast<int>(m_Player[1]->GetPos().x) + m_Player[1]->GetSizeLeft();
+	int PlayerTop = static_cast<int>(m_Player[1]->GetPos().y) + m_Player[1]->GetSizeTop();
+	int PlayerRight = static_cast<int>(m_Player[1]->GetPos().x) + m_Player[1]->GetSizeRight();
+	int PlayerBottom = static_cast<int>(m_Player[1]->GetPos().y) + m_Player[1]->GetSizeBottom();
 
-	int PlayerAtkLeft = m_Player[0]->GetPos().x + m_Player[0]->GetAttackSizeLeft();
-	int PlayerAtkTop = m_Player[0]->GetPos().y + m_Player[0]->GetAttackSizeTop();
-	int PlayerAtkRight = m_Player[0]->GetPos().x + m_Player[0]->GetAttackSizeRight();
-	int PlayerAtkBottom = m_Player[0]->GetPos().y + m_Player[0]->GetAttackSizeBottom();
+	int PlayerAtkLeft = static_cast<int>(m_Player[0]->GetPos().x) + m_Player[0]->GetAttackSizeLeft();
+	int PlayerAtkTop = static_cast<int>(m_Player[0]->GetPos().y) + m_Player[0]->GetAttackSizeTop();
+	int PlayerAtkRight = static_cast<int>(m_Player[0]->GetPos().x) + m_Player[0]->GetAttackSizeRight();
+	int PlayerAtkBottom = static_cast<int>(m_Player[0]->GetPos().y) + m_Player[0]->GetAttackSizeBottom();
 
 	if (PlayerLeft > PlayerAtkRight) return false;
 	if (PlayerRight < PlayerAtkLeft) return false;
